+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - added Args class to pass args to call setter
 - added inline binding
 - execute method from Expression is changed to function and can be used as dependency
+- "compilation" package is renamed to "expression"
 
 ## 2.3.0